feat-#24: move hotel repository to db module
