+++ src/catalog/hotel/commands/handlers/update-hotel.handler.ts
@@ -2,7 +2,7 @@ import { CommandHandler, ICommandHandler } from '@nestjs/cqrs';
 import { InjectRepository } from '@nestjs/typeorm';
 
 import { RpcExceptionService } from '../../../../utils/exception-handling';
-import { HotelRepository } from '../../repositories';
+import { HotelRepository } from '../../../../db/repositories';
 import { UpdateHotelCommand } from '../impl';
 import { HotelEntity } from '../../../../db/entities/hotel.entity';
