--- src/catalog/hotel/repositories/index.ts
@@ -1,1 +1,0 @@
-export * from './hotel.repository';
