+++ src/db/repositories/hotel.repository.ts
@@ -1,6 +1,6 @@
 import { EntityRepository, Repository } from 'typeorm';
 
-import { HotelEntity } from '../../../db/entities/hotel.entity';
+import { HotelEntity } from '../entities/hotel.entity';
 
 @EntityRepository(HotelEntity)
 export class HotelRepository extends Repository<HotelEntity> {}
